Changes from rebase of util -> protocol rename
