+++ org-code-javabuilder/neighborhood/build.gradle
@@ -23,7 +23,7 @@ dependencies {
     testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.6.0'
     // https://mvnrepository.com/artifact/org.json/json
     implementation group: 'org.json', name: 'json', version: '20210307'
-    implementation project(':util')
+    implementation project(':protocol')
 }
 
 // Builds the Neighborhood fat-jar (i.e. jar with all dependencies) and copies it to the resources
